[RES] Release 0.64

+++ utils/archive-source.sh
@@ -9,6 +9,8 @@
 # This just does all of the things necessary to prepare an archive (zipfile)
 # of the Visopsys sources and utilities.
 
+ZIPLOG=./zip.log
+
 echo ""
 echo "Making Visopsys SOURCE archive"
 echo ""
@@ -17,15 +19,15 @@ echo ""
 # the release number in the destination directory name.  Otherwise, we
 # assume an interim package and use the date instead
 if [ "$1" = "-r" ] ; then
-    # What is the current release version?
-    RELEASE=`./release.sh`
-    echo "(doing RELEASE version $RELEASE)"
-    echo ""
+	# What is the current release version?
+	RELEASE=`./release.sh`
+	echo "(doing RELEASE version $RELEASE)"
+	echo ""
 else
-    # What is the date?
-    RELEASE=`date +%Y-%m-%d`
-    echo "(doing INTERIM version $RELEASE -- use -r flag for RELEASES)"
-    echo ""
+	# What is the date?
+	RELEASE=`date +%Y-%m-%d`
+	echo "(doing INTERIM version $RELEASE -- use -r flag for RELEASES)"
+	echo ""
 fi
 
 DESTDIR=visopsys-"$RELEASE"-src
@@ -40,7 +42,6 @@ mv /tmp/"$DESTDIR" ./
 # Make sure it's clean
 echo -n "Making clean... "
 make -C "$DESTDIR" clean >& /dev/null
-
 # Remove all the things we don't want to distribute
 # CVS droppings
 find "$DESTDIR" -name CVS -exec rm -R {} \; >& /dev/null
@@ -48,23 +49,27 @@ find "$DESTDIR" -name CVS -exec rm -R {} \; >& /dev/null
 rm -Rf "$DESTDIR"/work
 rm -f "$DESTDIR"/src/ISSUES.txt
 rm -Rf "$DESTDIR"/src/patches
+echo Done
 
-echo done
 echo -n "Archiving... "
-
 echo "Visopsys $RELEASE Source Release" > /tmp/comment
-echo "Copyright (C) 1998-2006 J. Andrew McLaughlin" >> /tmp/comment    
+echo "Copyright (C) 1998-2006 J. Andrew McLaughlin" >> /tmp/comment
 rm -f "$DESTDIR".zip
-zip -9 -z -r "$DESTDIR".zip "$DESTDIR" < /tmp/comment >& /dev/null
+zip -9 -z -r "$DESTDIR".zip "$DESTDIR" < /tmp/comment >& $ZIPLOG
+if [ $? -ne 0 ] ; then
+	echo ""
+	echo -n "Not able to create zip file "$DESTDIR".zip.  "
+	echo "See $ZIPLOG.  Terminating."
+	echo ""
+	exit 1
+fi
+rm -f /tmp/comment $ZIPLOG
+echo Done
 
-echo done
 echo -n "Cleaning up... "
-
 # Remove the working directory
 rm -Rf "$DESTDIR"
-rm /tmp/comment
-
-echo done
+echo Done
 
 echo "File is: $DESTDIR.zip"
 echo ""
